create add update and delete finished

+++ lib/ORM/Manager.php
@@ -26,17 +26,14 @@ class Manager {
 
     public function find($id) {
 
-        $query = sprintf("SELECT * FROM %s WHERE id= %s",$this->datastructure["table"],$id);
+        $query = sprintf("SELECT * FROM %s WHERE id= %s", $this->datastructure["table"], $id);
         $requete = $this->pdo->prepare($query);
-        $requete->execute(array(":table"=> $this->datastructure["table"],":id" => $id));        
-        $result =  $requete->fetch(\PDO::FETCH_ASSOC);
+        $requete->execute(array(":table" => $this->datastructure["table"], ":id" => $id));
+        $result = $requete->fetch(\PDO::FETCH_ASSOC);
         $instance = new $this->entity();
-        $proper = $this->verifPropertyofEntity($result);
-        var_dump($proper);
-        var_dump($instance->hydrate($result));
-        exit();
-        $instance->hydrate($result);
-//        return (new $this->entity())->hydrate($result);
+
+        $instance->hydrate($result, $this->datastructure);
+        return (new $this->entity())->hydrate($result);
     }
 
     public function add(Entity $entity) {
@@ -54,13 +51,11 @@ class Manager {
     }
 
     public function modify(Entity $entity) {
- 
         $set = [];
         $parasm = [];
         foreach ($this->datastructure['columns'] as $col) {
             $property = $col['property'];
             $proper = $this->verifProperty($property);
-            var_dump($proper);
 
             if (!empty($entity->$property)) {
 
@@ -68,43 +63,41 @@ class Manager {
                 $parasm[$col['property']] = $entity->$property;
             }
         }
-        var_dump($entity);
-        var_dump($set);
-        var_dump($parasm);
-        var_dump(count($set));
 
         if (count($set)) {
-            $query = sprintf("UPDATE :s SET :s WHERE :s = :id", $entity::dataStructure()["table"], implode(",", $set), $entity::dataStructure()["primaryKey"]);
+            $query = sprintf("UPDATE %s SET %s WHERE %s = :id", $entity::dataStructure()["table"], implode(",", $set), $entity::dataStructure()["primaryKey"]);
             $requete = $this->pdo->prepare($query);
-            var_dump($query);
-//            exit();
             $requete->execute($parasm);
+
             return true;
         }
         return false;
     }
 
     public function delete(Entity $entity) {
-        $query = sprintf("DELETE FROM %s WHERE %s = :id", $entity::dataStructure()["table"], $entity::dataStructure()["primaryKey"]);
-        $requete = $this->pdo->prepare($sqlQuery);
-        $requete->execute(["id" => $entity::dataStructure()["primaryKey"]]);
-    }
-    
-    public function verifPropertyofEntity($property) {
-        $this->datastructure["columns"][$key] = $property;
 
-            
+        if (!isset($entity->id)) {
+            throw new \InvalidArgumentException("erreur : il n'y a aucun identifiant");
+        }
+
+        $query = sprintf("DELETE FROM %s WHERE %s = :id", $entity::dataStructure()["table"], $entity::dataStructure()["primaryKey"]);
+        $requete = $this->pdo->prepare($query);
+        $requete->execute([$entity::dataStructure()["primaryKey"] => $entity->id]);
+        
+        return true;
     }
-//    public function verifProperty($property) {
-//        if (preg_match('#^[^A-Z]*([A-Z].*)#', $property, $result)) {
-//            $result1 = str_replace($result[1], "", $property);
-////            
-//            $result2 = lcfirst($result[1]);
-//            $result = $result1 . "_" . $result2;
-//            return $result;
-//        }
-//        return $property;
-//    }
+
+    public function verifProperty($property) {
+        if (preg_match('#^[^A-Z]*([A-Z].*)#', $property, $result)) {
+            $result1 = str_replace($result[1], "", $property);
+//            
+            $result2 = lcfirst($result[1]);
+            $result = $result1 . "_" . $result2;
+            return $result;
+        }
+        return $property;
+    }
+
 //    public function getList($debut = -1, $limite = -1) {
 ////        var_dump($this->dataStructure);
 ////        foreach ($this->dataStructure as $key => $value) {
@@ -132,5 +125,4 @@ class Manager {
 //        return $this->pdo->query('SELECT COUNT(*) FROM article')->fetchColumn();
 //    }
 //
-
 }
